Move Broadway callbacks to Broadway.Server
Plus since we only need the supervisor pid,
that's the only thing we will keep as state.

Another change is that we unrolled the reduce
loops into for-comprehensions for simplicity.
Sometimes we have to traverse the list twice
but performance is not a concern here.

Finally, we build all childspecs by hand, since
we need to pass the id in anyway.