Rename batcher_key to batcher and partition to batch_key (#56)
Because we may have multiple partitions along the way,
it is probably best to distance ourselves from the
"partition" name. Instead, `batch_key` is used, as it
is one of the factor for specifying a batch beyond the
batch_size and batch_timeout.

To avoid  confusion, batcher_key in `%BatchInfo{}` was
rename to batcher (to mirror `put_batcher`) and `batcher_pid`
was removed for now.