Add "keys" to telemetry events and topology/1
This commit adds `batcher_key` and `processor_key` to events that don't
have it yet. This attribute is the same from the name of the processors
or the batchers that goes in the Broadway configuration.

The same keys are present now in the topology description. This can help
users to match telemetry events with the topology.