fix: documentation example to configure BroadwayEctoSandbox
It fixes the next telemetry warning:

```
Function passed as a handler with ID {App.BroadwayEctoSandbox, App.Repo} is local function.
This mean that it is either anonymous function or capture of function without module specified. That may cause performance penalty when calling such handler. For more details see note in `telemetry:attach/4` documentation.

https://hexdocs.pm/telemetry/telemetry.html#attach-4
```